UCMS: pattern updated LinkGrabber: weiteres flag ob linkgrabber läuft, da das adden im thread abläuft LinkGrabber: bug im packagecombine beseitigt DownloadController:optimierung auskommentiert da noch nicht fertig
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@5735 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/host/Netloadin.java
@@ -422,7 +422,7 @@ public class Netloadin extends PluginForHost {
             if (entries == null) {
                 entries = br.getRegex(";(.*?);(.*?);(.*?)").getRow(0);
                 if (entries == null || entries.length < 3) throw new PluginException(LinkStatus.ERROR_FILE_NOT_FOUND);
-                downloadLink.setDownloadSize((int) Regex.getSize(entries[1] + " bytes"));
+                downloadLink.setDownloadSize(Regex.getSize(entries[1] + " bytes"));
                 downloadLink.setFinalFileName(entries[0]);
                 fileStatusText = JDLocale.L("plugins.hoster.netloadin.errors.mightbeoffline", "Might be offline");
                 downloadLink.getLinkStatus().setStatusText(fileStatusText);
@@ -433,7 +433,7 @@ public class Netloadin extends PluginForHost {
 
             downloadLink.setFinalFileName(entries[1]);
             fileStatusText = entries[2];
-            downloadLink.setDownloadSize((int) Regex.getSize(entries[2] + " bytes"));
+            downloadLink.setDownloadSize(Regex.getSize(entries[2] + " bytes"));
 
             downloadLink.setMD5Hash(entries[4].trim());
             if (entries[3].equalsIgnoreCase("online")) { return AvailableStatus.TRUE; }
